academic_tool/requirements.txt
flask>=2.0.0
dashscope>=1.1.0
Pillow>=9.0.0
werkzeug>=2.0.0